Join Game and Game Start
Once a user is registered, they are able to join a game.
The game must not be full, there is a maximum of 4 players for all games.
This functionality must make use of synchronicity to provide accurate information when a user attempts to join

When four players have joined a game a game may be started and all players must be redirected
to the Game page where a new game is started.
If four players have not joined, the game starting functionality is not available.

Tasks:
https://github.com/pseguin2011/dame_de_pique/projects/1#card-47738104
https://github.com/pseguin2011/dame_de_pique/projects/1#card-47738456